requirements.txt
# ─────────────────────────────────────────────
#  PlagioAI — Python Requirements
# ─────────────────────────────────────────────

# Web Framework
flask>=3.0.0
flask-cors>=4.0.0
werkzeug>=3.0.0

# Document Parsing
pdfplumber>=0.10.0
python-docx>=1.1.0

# NLP
nltk>=3.8.0
sentence-transformers>=2.7.0
scikit-learn>=1.4.0

# AI Rewriting
transformers>=4.40.0
torch>=2.2.0          # CPU version — for GPU use: torch>=2.2.0+cu121

# Report Generation
reportlab>=4.1.0

# Utilities
numpy>=1.26.0
